pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "chatdku-mini-rag"
version = "0.1.0"
description = "Mini bilingual agentic RAG pipeline for the ChatDKU candidate task."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pypdf>=4.0.0",
  "python-docx>=1.1.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
